chore(insights): inline clustering helpers and drop new unit tests
Trims the annotation overlay PR under the 500-line review target by
removing the useAnnotationClusters hook extraction and the new
annotationBadgeDataIndices selector tests. The clustering logic and
interpolation helper are re-inlined in AnnotationsOverlay.tsx with the
rightPx extent fix preserved.

The extraction + unit tests will land as a stacked follow-up PR.
