+++ frontend/src/lib/components/AnnotationsOverlay/annotationsOverlayLogic.test.ts
@@ -868,80 +868,4 @@ describe('annotationsOverlayLogic', () => {
             })
         })
     })
-
-    describe('annotationBadgeDataIndices', () => {
-        async function mountAndWait(props: Parameters<typeof annotationsOverlayLogic>[0]): Promise<void> {
-            logic = annotationsOverlayLogic(props)
-            logic.mount()
-            await expectLogic(annotationsModel).toDispatchActions(['loadAnnotationsSuccess'])
-            await expectLogic(
-                insightLogic({ dashboardItemId: MOCK_INSIGHT_SHORT_ID, dashboardId: MOCK_DASHBOARD_ID })
-            ).toDispatchActions(['loadInsightSuccess'])
-        }
-
-        function indicesByKey(
-            values: ReturnType<typeof annotationsOverlayLogic.build>['values']
-        ): Record<string, number> {
-            return Object.fromEntries(
-                values.annotationBadgeDataIndices.map(
-                    (b: { dateKey: string; dataIndex: number }) => [b.dateKey, b.dataIndex] as const
-                )
-            )
-        }
-
-        it.each<{
-            name: string
-            interval: IntervalType
-            dates: string[]
-            expected: Array<[string, number]>
-        }>([
-            {
-                name: 'monthly chart → fractional indices for mid-month annotations',
-                interval: 'month',
-                dates: ['2022-08-01', '2022-09-01', '2022-10-01'],
-                expected: [
-                    ['2022-08-10 00:00:00+0000', 9 / 31],
-                    ['2022-09-10 00:00:00+0000', 1 + 9 / 30],
-                ],
-            },
-            {
-                // Regression: startOf('week') drifts Monday-aligned dates back to Sunday.
-                name: 'weekly Monday-aligned chart → no Sunday drift',
-                interval: 'week',
-                dates: ['2022-08-08', '2022-08-15', '2022-08-22', '2022-08-29', '2022-09-05', '2022-09-12'],
-                expected: [
-                    ['2022-08-10 00:00:00+0000', 2 / 7],
-                    ['2022-08-17 00:00:00+0000', 1 + 2 / 7],
-                ],
-            },
-        ])('$name', async ({ interval, dates, expected }) => {
-            useInsightMocks(interval)
-            await mountAndWait({
-                dashboardItemId: MOCK_INSIGHT_SHORT_ID,
-                insightNumericId: MOCK_INSIGHT_NUMERIC_ID,
-                dates,
-                ticks: dates.map((_, i) => ({ value: i })),
-                dashboardId: MOCK_DASHBOARD_ID,
-            })
-
-            const indices = indicesByKey(logic.values)
-            for (const [dateKey, expectedIndex] of expected) {
-                expect(indices[dateKey]).toBeCloseTo(expectedIndex, 5)
-            }
-        })
-
-        it('clamps annotations past the last data point to the last index', async () => {
-            useInsightMocks('month')
-            await mountAndWait({
-                dashboardItemId: MOCK_INSIGHT_SHORT_ID,
-                insightNumericId: MOCK_INSIGHT_NUMERIC_ID,
-                dates: ['2022-08-01', '2022-09-01'],
-                ticks: [{ value: 0 }, { value: 1 }],
-                dashboardId: MOCK_DASHBOARD_ID,
-            })
-
-            // Sep 10 would extrapolate to ~1.3 without clamping; must snap to lastIndex (1).
-            expect(indicesByKey(logic.values)['2022-09-10 00:00:00+0000']).toBe(1)
-        })
-    })
 })
